matching_agent/requirements.txt
google-adk
python-dotenv
pydantic
pydantic-settings